- Reduced visibility of BridgeStudyStaffClient (like other implementations, it's hidden to end users); - Made ResourceList concrete as PagedResourceList, and moved it to the models package; - combined users package into the account package.
All these changes make the javadocs of public classes/methods clearer and simpler.
