+++ java-sdk/src/main/java/org/sagebionetworks/bridge/sdk/BridgeSession.java
@@ -5,10 +5,10 @@ import static com.google.common.base.Preconditions.checkState;
 
 import java.util.Map;
 
+import org.sagebionetworks.bridge.sdk.models.accounts.SharingScope;
 import org.sagebionetworks.bridge.sdk.models.accounts.StudyParticipant;
 import org.sagebionetworks.bridge.sdk.models.subpopulations.ConsentStatus;
 import org.sagebionetworks.bridge.sdk.models.subpopulations.SubpopulationGuid;
-import org.sagebionetworks.bridge.sdk.models.users.SharingScope;
 
 import com.google.common.collect.ImmutableMap;
