+++ java-sdk/src/main/java/org/sagebionetworks/bridge/sdk/BridgeStudyStaffClient.java
@@ -2,7 +2,7 @@ package org.sagebionetworks.bridge.sdk;
 
 import org.sagebionetworks.bridge.sdk.models.studies.Study;
 
-public class BridgeStudyStaffClient extends BaseApiCaller implements StudyStaffClient {
+class BridgeStudyStaffClient extends BaseApiCaller implements StudyStaffClient {
     
     BridgeStudyStaffClient(BridgeSession session) {
         super(session);
