+++ java-sdk/src/main/java/org/sagebionetworks/bridge/sdk/models/ResourceList.java
@@ -1,11 +1,73 @@
 package org.sagebionetworks.bridge.sdk.models;
 
+import static com.google.common.base.Preconditions.checkElementIndex;
+
 import java.util.Iterator;
 import java.util.List;
 
-public interface ResourceList<T> extends Iterable<T> {
-    public List<T> getItems();
-    public int getTotal();
-    public T get(int index);
-    @Override public Iterator<T> iterator();
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public final class ResourceList<T> {
+
+    private final List<T> items;
+    private final int total;
+    
+    @JsonCreator
+    public ResourceList(@JsonProperty("items") List<T> items, @JsonProperty("total") int total) {
+        this.items = items;
+        this.total = total;
+    }
+    
+    public List<T> getItems() {
+        return items;
+    }
+    
+    public int getTotal() {
+        return total;
+    }
+
+    public T get(int index) {
+        checkElementIndex(index, items.size());
+        return items.get(index);
+    }
+    
+    public Iterator<T> iterator() {
+        return items.iterator();
+    }
+
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + total;
+        result = prime * result + ((items == null) ? 0 : items.hashCode());
+        return result;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj)
+            return true;
+        if (obj == null)
+            return false;
+        if (getClass() != obj.getClass())
+            return false;
+        @SuppressWarnings("rawtypes")
+        ResourceList other = (ResourceList) obj;
+        if (total != other.total)
+            return false;
+        if (items == null) {
+            if (other.items != null)
+                return false;
+        } else if (!items.equals(other.items))
+            return false;
+        return true;
+    }
+
+    @Override
+    public String toString() {
+        return "ResourceList [items=" + items + ", total=" + total + "]";
+    }
+    
 }
