+++ java-sdk/src/main/java/org/sagebionetworks/bridge/sdk/models/accounts/SignInCredentials.java
@@ -1,4 +1,4 @@
-package org.sagebionetworks.bridge.sdk.models.users;
+package org.sagebionetworks.bridge.sdk.models.accounts;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
